wif_Django/pyproject.toml
[tool.poetry]
name = "wif_django"
version = "0.1.0"
description = ""
authors = ["Hordunlarmy <Hordunlarmy@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^5.0.6"
python-decouple = "^3.8"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
